Added 2 commands
Dm
SetAmongUsRole

The commands speak for themself.

The Dm command only works for Bot Moderators
SetAmongUsRole works for everyone, I've to fix this next commit.

The GuildDB now also gets added to the event.

You can now get the first role mentioned using e.getFirstRoleMentioned

Fixed a bug DatabaseGuild that resulted in info not getting updated

Added amongUsRoleId to GuildDB

Now you can set a among us role, so when using !startgame it will mention that role. Or it won't mention any roles

When doing !level it also show the reps

Fixed icon of !startgame

Want to add todolist to a UserDB, but this is mostly speculating.