Only people with the manage role permission can set the among us role.
The ID of a joke is now a String instead int

Added !setblacklist as alias for !setblacklisted

Fixed description of RemoveJoke command

It sends the joke one last time after removing it.
If you don't want it to send the joke you can put false as second argument for the command.