Beta to master (#3)
* Getting the current color is now static

* Added bot moderator system.

Only certain users can run a few commands

The userDB item of the user is now added to the event.
The UserDB contains the user his level, xp, userID and or he's a moderator.

* Added 2 commands

SetBotModerator
SetBlacklisted

The commands speak for themself.
Working on bot moderation category (so category will be invisible except for bot moderators)

Someone's UserDB object is now added to the event (to check or user is blacklisted/ a bot moderator, thought I could just add it to the event)

Get the first user mentioned in a message using
e.getFirstUserMentioned

* Added bot moderation category, only bot moderators will see the help message and can run the commands.

Now before running the command it already checks or a user is a bot moderator.

* Added reps (rep a user using !rep see reps using !rep (user Id)

Shutdown command has been fixed and now is in the botmoderation category, and doesn't check or the user is a moderator

Start of caching, command delay has been lowered a lot since it just gets the user from cache, not from the database.

* Added 2 commands
Avatar (get a user's or your avatar)
Say (say something in specific channels (bot moderators only))

Get the first channel mentioned using e.getFirstChannelMentioned

Rep command now has a cooldown

Added function to get the first arg as user (in event)

Slight change in StartGame

Optimalizations:

Guilds now also get cached
Guild handler small optimalization
Removes a lot of code from Level command
Removed object from ServerStats so it now uses a static method

* Added 2 commands

Dm
SetAmongUsRole

The commands speak for themself.

The Dm command only works for Bot Moderators
SetAmongUsRole works for everyone, I've to fix this next commit.

The GuildDB now also gets added to the event.

You can now get the first role mentioned using e.getFirstRoleMentioned

Fixed a bug DatabaseGuild that resulted in info not getting updated

Added amongUsRoleId to GuildDB

Now you can set a among us role, so when using !startgame it will mention that role. Or it won't mention any roles

When doing !level it also show the reps

Fixed icon of !startgame

Want to add todolist to a UserDB, but this is mostly speculating.

* Only people with the manage role permission can set the among us role.

The ID of a joke is now a String instead int

Added !setblacklist as alias for !setblacklisted

Fixed description of RemoveJoke command

It sends the joke one last time after removing it.
If you don't want it to send the joke you can put false as second argument for the command.