Start of Github Pages, bug fixes and command updates.
Trying to get a Github Pages website.  botPrefix command is now static, so it doesn't look like garbage.  Fixed a bug that it made the CommandReceivedEvent twice, and checking or the message starts with the correct prefix now works properly in DM's.  Whois/userinfo command now also shows the voicestate (whenever a member is in a voice channel or not)  Tried to get rich presence working again, but found it I can't set it to streaming.